Proper capitalisation on nouns, pronouns; Twitter defaults to HTTPS
December and GitHub should be spelt with the proper casing of the first
letter for both, the humpy H in the case of the latter.

Twitter redirects all links to HTTPS version, clean it up a little by
just showing the twitter.com without the leading http since people
rarely enter that part of a URL any more.