Move "Host Event" button to bottom of events list
With a very short list of events, the map takes up too much room and
moves the button “below the fold”. It was unclear to me exactly how to
add a new event. By putting it at the bottom of the list, it’s in a
more logical location I think.